LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-003-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 2 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 3 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 4 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 5 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 6 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
FAIL .target 7 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 0
